+++ src/legacy/client__shell__optsback.h
@@ -3,7 +3,7 @@
 
 #include "defs.h"
 
-int sub_4A18E0(void* a1, int a2, int a3, int a4);
+uintptr_t sub_4A18E0(void* a1, int a2, uintptr_t a3, uintptr_t a4);
 void sub_4A19F0(char* a1);
 
 #endif // NOX_PORT_CLIENT_SHELL_OPTSBACK
