+++ src/legacy/gui_console.go
@@ -16,13 +16,13 @@ var (
 )
 
 //export nox_gui_console_flagXxx_451410
-func nox_gui_console_flagXxx_451410() int { return Nox_gui_console_flagXxx_451410() }
+func nox_gui_console_flagXxx_451410() int32 { return int32(Nox_gui_console_flagXxx_451410()) }
 
 //export nox_gui_console_Print_450B90
-func nox_gui_console_Print_450B90(cl C.uchar, cstr *wchar2_t) int {
+func nox_gui_console_Print_450B90(cl C.uchar, cstr *wchar2_t) int32 {
 	// since legacy code still calls it, we redirect into global printer instead of GUI printer
 	GetConsole().Print(console.Color(cl), GoWString(cstr))
-	return 1
+	return int32(1)
 }
 
 //export nox_gui_console_PrintOrError_450C30
@@ -31,7 +31,7 @@ func nox_gui_console_PrintOrError_450C30(cl C.uchar, cstr *wchar2_t) {
 }
 
 //export nox_gui_console_Hide_4512B0
-func nox_gui_console_Hide_4512B0() int { return Nox_gui_console_Hide_4512B0() }
+func nox_gui_console_Hide_4512B0() int32 { return int32(Nox_gui_console_Hide_4512B0()) }
 
 func Nox_xxx_consoleEsc_49B7A0() {
 	C.nox_xxx_consoleEsc_49B7A0()
