+++ src/legacy/gui_noxworld.go
@@ -188,15 +188,15 @@ func nox_client_refreshServerList_4378B0() { Nox_client_refreshServerList_4378B0
 func sub_438770_waitList() { Sub_438770_waitList() }
 
 //export nox_xxx_createSocketLocal_554B40
-func nox_xxx_createSocketLocal_554B40(port uint16) int {
+func nox_xxx_createSocketLocal_554B40(port uint16) int32 {
 	if err := Nox_xxx_createSocketLocal(int(port)); err != nil {
-		return -1
+		return int32(-1)
 	}
-	return 0
+	return int32(0)
 }
 
 //export sub_554D10
-func sub_554D10() int { return Sub_554D10() }
+func sub_554D10() int32 { return int32(Sub_554D10()) }
 
 //export sub_4373A0
 func sub_4373A0() { Sub_4373A0() }
