+++ src/legacy/gui_summon.go
@@ -14,4 +14,4 @@ var (
 )
 
 //export sub_4C26F0
-func sub_4C26F0(a1 unsafe.Pointer) int { return Sub_4C26F0(AsWindowP(a1)) }
+func sub_4C26F0(a1 unsafe.Pointer) int32 { return int32(Sub_4C26F0(AsWindowP(a1))) }
